Disable "pull" when building containers in CI
The "pull-always" has changed in podman 3.2, causing some
headache for us.

However, in our build containers, we do not have a cache for
podman anyhow, thus we should probably not be using
"pull-always" anyhow. It's enough to do explicit pulls if need
be.

This overrides the build.mk default of "pull-always" to a more
conservative "pull if missing".

Issue: ModioAB/agile#2252
Upstream issue: https://github.com/containers/podman/issues/10307