Add shared logic for rust include code
This sets up the logic I think is useful for our embedded / cross
compiled project to a target.

None of the rules are enabled by default, as to not cause surprises for
those that include it.

Some of them require a PACKAGENAME to be set as a variable before they
can be used, as that is the rust package that should be built/created.

There are no needs or dependencies, as I was expecting that to be used
elsewhere, and while it does declare many of the most commonly used
things, this isn't rule of law.